# Conflicts:
#	ebegu-server/src/main/java/ch/dvbern/ebegu/rechner/KitaRechner.java
#	ebegu-server/src/main/java/ch/dvbern/ebegu/rechner/TagiRechner.java
#	ebegu-shared/src/main/java/ch/dvbern/ebegu/entities/Betreuungspensum.java

+++ ebegu-server/src/test/java/ch/dvbern/ebegu/rechner/AbstractBGRechnerTest.java
@@ -253,7 +253,7 @@ public class AbstractBGRechnerTest {
 
 		VerfuegungZeitabschnitt zeitabschnitt = new VerfuegungZeitabschnitt(new DateRange(von, bis));
 		zeitabschnitt.setAnspruchberechtigtesPensum(anspruch);
-		zeitabschnitt.setBetreuungspensum(anspruch);
+		zeitabschnitt.setBetreuungspensum(BigDecimal.valueOf(anspruch));
 		zeitabschnitt.setMassgebendesEinkommenVorAbzugFamgr(massgebendesEinkommen);
 		zeitabschnitt.setMonatlicheBetreuungskosten(monatlicheBetreuungskosten);
 		List<VerfuegungZeitabschnitt> zeitabschnittList = new ArrayList<>();
