+++ ebegu-testutils/src/main/java/ch/dvbern/ebegu/test/TestDataUtil.java
@@ -676,7 +676,7 @@ public final class TestDataUtil {
 
 	private static Betreuungspensum createBetreuungspensum() {
 		Betreuungspensum betreuungspensum = new Betreuungspensum();
-		betreuungspensum.setPensum(80);
+		betreuungspensum.setPensum(BigDecimal.valueOf(80));
 		return betreuungspensum;
 	}
 
@@ -1258,7 +1258,8 @@ public final class TestDataUtil {
 
 	public static void createDefaultAdressenForGS(final Gesuch gesuch, final boolean gs2) {
 		List<GesuchstellerAdresseContainer> adressen1 = new ArrayList<>();
-		final GesuchstellerAdresseContainer adresseGS1 = TestDataUtil.createDefaultGesuchstellerAdresseContainer(gesuch.getGesuchsteller1());
+		final GesuchstellerAdresseContainer adresseGS1 = TestDataUtil.
+			createDefaultGesuchstellerAdresseContainer(gesuch.getGesuchsteller1());
 		Objects.requireNonNull(adresseGS1.getGesuchstellerAdresseJA());
 		adresseGS1.getGesuchstellerAdresseJA().setNichtInGemeinde(false);
 		adresseGS1.getGesuchstellerAdresseJA().setGueltigkeit(new DateRange(Constants.START_OF_TIME, Constants.END_OF_TIME));
@@ -1268,7 +1269,8 @@ public final class TestDataUtil {
 
 		if (gs2) {
 			List<GesuchstellerAdresseContainer> adressen2 = new ArrayList<>();
-			final GesuchstellerAdresseContainer adresseGS2 = TestDataUtil.createDefaultGesuchstellerAdresseContainer(gesuch.getGesuchsteller2());
+			final GesuchstellerAdresseContainer adresseGS2 = TestDataUtil
+				.createDefaultGesuchstellerAdresseContainer(gesuch.getGesuchsteller2());
 			Objects.requireNonNull(adresseGS2.getGesuchstellerAdresseJA());
 			adresseGS2.getGesuchstellerAdresseJA().setNichtInGemeinde(false);
 			adresseGS2.getGesuchstellerAdresseJA().setGueltigkeit(new DateRange(Constants.START_OF_TIME, Constants.END_OF_TIME));
@@ -1515,7 +1517,7 @@ public final class TestDataUtil {
 	public static VerfuegungZeitabschnitt createDefaultZeitabschnitt(Verfuegung verfuegung) {
 		VerfuegungZeitabschnitt zeitabschnitt = new VerfuegungZeitabschnitt();
 		zeitabschnitt.setVerfuegung(verfuegung);
-		zeitabschnitt.setBetreuungspensum(10);
+		zeitabschnitt.setBetreuungspensum(BigDecimal.valueOf(10));
 		zeitabschnitt.setAnspruchberechtigtesPensum(50);
 		zeitabschnitt.setEinkommensjahr(PERIODE_JAHR_1);
 		zeitabschnitt.setZuSpaetEingereicht(false);
