+++ ebegu-testutils/src/main/java/ch/dvbern/ebegu/test/data/VerfuegungZeitabschnittData.java
@@ -31,7 +31,7 @@ public class VerfuegungZeitabschnittData {
 
 	private String gueltigBis;
 
-	private int betreuungspensum;
+	private BigDecimal betreuungspensum;
 
 	private int anspruchberechtigtesPensum;
 
@@ -63,11 +63,11 @@ public class VerfuegungZeitabschnittData {
 		this.vollkosten = verfuegungZeitabschnitt.getVollkosten();
 	}
 
-	public int getBetreuungspensum() {
+	public BigDecimal getBetreuungspensum() {
 		return betreuungspensum;
 	}
 
-	public void setBetreuungspensum(int betreuungspensum) {
+	public void setBetreuungspensum(BigDecimal betreuungspensum) {
 		this.betreuungspensum = betreuungspensum;
 	}
